HPC: Remove MPI-single machine tests
MPI running on a single machine is not a real-life scenario. In the
past there was some bugs reported but still such bugs have little
real-life value. It is better to focus on multimachine scenarios and
this patch is doing exactly this - removing single-machine executions
and leaves only multi-machine ones