kselftests: Update the tap output with known issues
Kernel selftests do not allow for individual test runs. Basic test
structure is: suite:test where test is a binary which might contain
still many single tests. For instance: cgroup:test_memcontrol contains
14 individual tests whith 3 tend to fail. For this reason it is pretty
much not possible now, to exclude single test from the execution. Known
issues from the LTP lib is being used for flagging the known issues,
maintaining the listh of known issues, but then only in the post
process the tap output can be changed to mark the failed test as
known issue and marked as softfailed. The TAP language does not know
softfail, so tests reported as "not ok" but found in the known_issues
are marked as "ok # TODO" and should be then marked as softfailed by the
KTAP parser